Intergrate More commonly used property in Frame Interface; Fix startup problem in eval_large; make idxFile a property (remove dependency on index file to be exist);  minor type fix

+++ LogInterface/TypeInfoChunk.py
@@ -190,7 +190,7 @@ class TypeInfoChunk(Chunk):
 
         size = sutil.readUInt()
         needsTypenameUnification = unifiedTypeNames & size == 0
-        size = size & ~unifiedTypeNames
+        size = UInt(np.int64(size) & ~np.int64(unifiedTypeNames))
         # print(f"Primitives num: {size}")
         for _ in range(size, 0, -1):
             type = sutil.readStr()
